test: improve FilterHelper test coverage
